added customisation for custom widget in create post, universal feed and post detail screen.
