+++ likeminds-feed-reactnative-integration/context/createPostContext.tsx
@@ -168,6 +168,12 @@ export interface CreatePostContextValues {
   ) => void;
   handleScreenBackPress: () => void;
   handleHeadingInputChange: (event: string) => void;
+  setDisabledTopicsGlobal: any;
+  disbaledTopicsGlobal: any;
+  showTopics: boolean;
+  setShowTopics: Dispatch<SetStateAction<boolean>>;
+  mappedTopics: any;
+  setMappedTopics: any;
 }
 
 const CreatePostContext = createContext<CreatePostContextValues | undefined>(
@@ -222,6 +228,9 @@ export const CreatePostContextProvider = ({
   const [allTags, setAllTags] = useState<Array<LMUserViewData>>([]);
   const [isUserTagging, setIsUserTagging] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
+  const [disbaledTopicsGlobal, setDisabledTopicsGlobal] = useState([] as any);
+  const [showTopics, setShowTopics] = useState(false);
+  const [mappedTopics, setMappedTopics] = useState([] as any);
 
   const maxHeadingWords = STYLES?.$CREATE_POST_STYLE?.headingMaxWords
     ? STYLES?.$CREATE_POST_STYLE?.headingMaxWords
@@ -833,6 +842,9 @@ export const CreatePostContextProvider = ({
     isUserTagging,
     isLoading,
     heading,
+    disbaledTopicsGlobal,
+    showTopics,
+    mappedTopics,
     setIsLoading,
     setIsUserTagging,
     setAllTags,
@@ -869,6 +881,9 @@ export const CreatePostContextProvider = ({
     handleLoadMore,
     onPostClick,
     handleScreenBackPress,
+    setDisabledTopicsGlobal,
+    setShowTopics,
+    setMappedTopics,
   };
 
   return (
